Refactor and move client package to internal directory
Move `client.go` and `client_test.go` to `internal/websocket` and update package and import paths accordingly. This change reorganizes the project structure to enhance maintainability and module encapsulation. Adjusted the references to configuration and model packages to match their new locations.
